requirements.txt
matplotlib==3.8.4
numpy==2.0.0
pandas==2.2.2
Pillow==10.4.0
torch==2.0.1
torchvision==0.15.2
